Add waffle's build rework (tnx waffle!)

+++ project/Aki.Build/Build.ps1
@@ -1,9 +1,4 @@
-﻿# VSCode.ps1
-# Copyright: © SPT-AKI 2020
-# License: NCSA
-# Authors:
-# - Merijn Hendriks
-# - waffle.lord
+﻿param ([switch]$VSBuilt)
 
 #copy and verify file function. Destination path should be a directory.
 function CopyAndVerifyFile 
@@ -62,35 +57,49 @@ function CopyAndVerifyFile
     Write-Host "Something went wrong :( `nError: $($Error[0])" -ForegroundColor Red
 }
 
-# locate msbuild
-Write-Host "Scanning for build tools..."
-
-$vsWhere = "${env:ProgramFiles(x86)}\Microsoft Visual Studio\Installer\vswhere.exe"
-
-if (($vsWhere -eq("")) -or(-not(Test-Path $vsWhere)))
+if(-not $VSBuilt)
 {
-    Write-Warning "  Could not find VSWhere.exe, please install BuildTools 2017 or newer"
-    return
+    # locate msbuild
+    Write-Host "Scanning for build tools..."
+
+    $vsWhere = "${env:ProgramFiles(x86)}\Microsoft Visual Studio\Installer\vswhere.exe"
+
+    if (($vsWhere -eq("")) -or(-not(Test-Path $vsWhere)))
+    {
+        Write-Warning "  Could not find VSWhere.exe, please install BuildTools 2017 or newer"
+        return
+    }
+
+    $msbuild = & $vsWhere -latest -products * -requires Microsoft.Component.MSBuild -find MSBuild\**\Bin\MSBuild.exe
+
+    if (($msbuild -eq("")) -or(-Not(Test-Path $msbuild)))
+    {
+        # make sure msbuild ins't empty and that the path exists, otherwise warn and exit.
+        Write-Warning "  Could not find Microsoft Buildtools"
+        return
+    }
+
+    Write-Host "  Found MSBuild.exe" -ForegroundColor Green
+    Write-Host ""
+
+    # restore nuget packages and build project
+    Write-Host "Building Modules..." -ForegroundColor Cyan
+    $buildProcess = Start-Process -FilePath $msbuild -NoNewWindow -ArgumentList "-nologo /verbosity:minimal -consoleloggerparameters:Summary -t:Restore;Rebuild -p:Configuration=Release Modules.sln" -PassThru
+    Wait-Process -InputObject $buildProcess
+    Write-Host "Done" -ForegroundColor Cyan
+    Write-Host ""
+}
+else
+{
+    Write-Host "VSBuilt: Skipping build" -ForegroundColor Cyan
 }
 
-$msbuild = & $vsWhere -latest -products * -requires Microsoft.Component.MSBuild -find MSBuild\**\Bin\MSBuild.exe
 
-if (($msbuild -eq("")) -or(-Not(Test-Path $msbuild)))
+if($VSBuilt) 
 {
-    # make sure msbuild ins't empty and that the path exists, otherwise warn and exit.
-    Write-Warning "  Could not find Microsoft Buildtools"
-    return
+    #move up a directory if the script is run from VS (it runs from the PostBuildProject dir)
+    cd ../../..
 }
-
-Write-Host "  Found MSBuild.exe" -ForegroundColor Green
-Write-Host ""
-
-# restore nuget packages and build project
-Write-Host "Building Modules..." -ForegroundColor Cyan
-$buildProcess = Start-Process -FilePath $msbuild -NoNewWindow -ArgumentList "-nologo /verbosity:minimal -consoleloggerparameters:Summary -t:Restore;Rebuild -p:Configuration=Release Modules.sln" -PassThru
-Wait-Process -InputObject $buildProcess
-Write-Host "Done" -ForegroundColor Cyan
-Write-Host ""
 
 #get root of project folder
 $rootPath = Resolve-Path -path "."
@@ -106,6 +115,21 @@ $akiModulesFolder = "$($buildDir)\Aki_Data\Modules"
 if(Test-Path $buildDir) 
 {
     Remove-Item $buildDir -Recurse -Force
+}
+
+$postBuildBin = "$($rootPath)\PostBuildProject\bin"
+$postBuildObj = "$($rootPath)\PostBuildProject\obj"
+
+if(Test-Path $postBuildBin)
+{
+    Remove-Item $postBuildBin -Recurse -Force
+    Write-Host "Removed PostBuildProject bin" -ForegroundColor Cyan
+}
+
+if(Test-Path $postBuildObj)
+{
+    Remove-Item $postBuildObj -Recurse -Force
+    Write-Host "Removed PostBuildProject obj" -ForegroundColor Cyan
 }
 
 #get all release dlls and exe files
@@ -134,15 +158,18 @@ foreach($file in $dllAndExeFiles)
 Write-Host ""
 
 # delete build waste
-Write-Host "Cleaning garbage produced by build..." -ForegroundColor Cyan
-
-$delPaths = Get-ChildItem -Recurse -Path $rootPath | where {$_.FullName -like "*\bin"} | select -ExpandProperty FullName
-$delPaths += Get-ChildItem -Recurse -Path $rootPath | where {$_.FullName -like "*\obj"} | select -ExpandProperty FullName
-
-foreach ($path in $delPaths)
+if(-not $VSBuilt)
 {
-    Write-Host "  Delete: $($path)"
-    Remove-Item $path -Force -Recurse
+    Write-Host "Cleaning garbage produced by build..." -ForegroundColor Cyan
+
+    $delPaths = Get-ChildItem -Recurse -Path $rootPath | where {$_.FullName -like "*\bin"} | select -ExpandProperty FullName
+    $delPaths += Get-ChildItem -Recurse -Path $rootPath | where {$_.FullName -like "*\obj"} | select -ExpandProperty FullName
+
+    foreach ($path in $delPaths)
+    {
+        Write-Host "  Delete: $($path)"
+        Remove-Item $path -Force -Recurse
+    }
 }
 
 Write-Host ""
